requirements.txt
# Core Django
Django==6.0.2
djangorestframework==3.15.2

# CORS Support
django-cors-headers==4.4.0

# Cloud Storage
cloudinary==1.41.0
django-cloudinary-storage==0.3.0

# Environment Variables
python-decouple==3.8

# Security
django-environ==0.11.2

# Development Tools
django-debug-toolbar==4.4.2
django-extensions==3.2.3

# Testing
factory-boy==3.3.0
pytest==8.2.2
pytest-django==4.8.0

# Documentation
drf-spectacular==0.27.2  # For OpenAPI/Swagger documentation

# Filtering
django-filter==24.3

# Production Server
gunicorn==21.2.0

# Static Files
whitenoise==6.6.0

# PostgreSQL Database
dj-database-url==2.1.0
psycopg[binary]==3.2.4

# JWT Authentication
djangorestframework-simplejwt==5.3.0

Pillow==11.1.0